pyproject.toml
[project]
name = "h1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "db-dtypes>=1.4.4",
    "google-auth>=2.43.0",
    "google-cloud-bigquery>=3.38.0",
    "pandas>=2.3.3",
]
